+++ R/affirm_report.R
@@ -5,12 +5,10 @@
 #' - `affirm_report_raw_data()` returns raw data used to generate summary in `affirm_report_gt()`
 #'
 #' @inheritParams openxlsx::write.xlsx
-#' @param variable_labels logical indicating whether to add a row to exported
-#' data with the variable labels. If label does not exist, the variable name
-#' is printed. Default is `FALSE`
-#' @param sheet_name A string for sheet names in the excel report; the item name
+#' @param affirmation_name A string for affirmation names; the item name
 #' in curly brackets is replaced with the item value (see glue::glue). Item names
 #' accepted include: `id`, `label`, `priority`, `data_frames`, `columns`, `error_n`, `total_n`.
+#' Defaults to `"{data_frames}{id}"`.
 #'
 #' @return gt table
 #' @name affirm_report
@@ -35,8 +33,8 @@ NULL
 
 #' @rdname affirm_report
 #' @export
-affirm_report_gt <- function(variable_labels = FALSE) {
-  affirm_report_raw_data(variable_labels = variable_labels) |>
+affirm_report_gt <- function() {
+  affirm_report_raw_data() |>
     dplyr::mutate(status_color = NA_character_, .before = 1L) |>
     dplyr::mutate(
       csv_download_link =
@@ -45,7 +43,6 @@ affirm_report_gt <- function(variable_labels = FALSE) {
           # these two args are the ones being passed to FUN
           .data$data,
           paste0("extract_", dplyr::row_number(), ".csv"),
-          variable_labels,
           # additional mapply args
           SIMPLIFY = TRUE,
           USE.NAMES = FALSE
@@ -59,42 +56,46 @@ affirm_report_gt <- function(variable_labels = FALSE) {
 
 #' @rdname affirm_report
 #' @export
-affirm_report_excel <- function(file, sheet_name = "{data_frames}{id}", variable_labels = FALSE, overwrite = TRUE) {
+affirm_report_excel <- function(file, affirmation_name = "{data_frames}{id}", overwrite = TRUE) {
 
-  # checking to make sure sheet name glue syntax has acceptable column names
-  sheet_name_cols <- regmatches(sheet_name, gregexpr("\\{([^\\}]+)\\}", sheet_name))[[1]] |>
-    gsub("\\{|\\}", "", x = _)
-
-  # acceptable variables to pass through glue syntax for sheet names
-  glue_accept <- c("id", "label", "priority", "data_frames", "columns", "error_n", "total_n")
-
-  # readable version for error messaging
-  glue_accept_str <- paste0("`", glue_accept, "`", collapse = ", ")
-  if (any(!sheet_name_cols %in% glue_accept)){
-    stop(paste0("`sheet_name` glue syntax expects one of ", glue_accept_str))
-  }
-
-  df_report <-
-    affirm_report_raw_data(variable_labels = variable_labels) |>
+  df_summary <-
+    affirm_report_raw_data() |>
     dplyr::filter(.data$error_n > 0L) |>
     dplyr::mutate(
-      label_final = glue::glue(sheet_name) |>
+      affirmation_name = glue::glue(affirmation_name) |>
         gsub(pattern = "[[:punct:]]", replacement = "", x = _)
-    )
+    ) |>
+    # change order of excel report
+    dplyr::select(
+      "affirmation_name", "data_frames", "id", "priority", "columns", "error_n",
+      "total_n", "error_rate", "label", "data"
+      )
 
   # checking to make sure sheet names are not too long
-  if (any(nchar(df_report$label_final) > 31)){
+  if (any(nchar(df_summary$affirmation_name) > 31)){
     stop("At least one sheet name exceeds the allowed 31 characters.")
   }
 
-  df_report$data |>
-    stats::setNames(df_report$label_final) |>
-    openxlsx::write.xlsx(file = file, overwrite = overwrite, colNames = !variable_labels)
+  # this is the affirmation data that gets exported to each sheets
+  # drops data column and columns with all NAs
+  df_export <- .identify_keep_data(df_summary) |>
+    # add empty Notes column to end
+    mutate(Notes = NA)
+
+  # create excel workbook with all affirmations
+  wb <- openxlsx2::wb_workbook() |>
+    .add_summary_sheet(df_export)
+
+  for (i in seq_len(nrow(df_summary))){
+    wb <- .add_affirmation_sheet(wb, df_summary[i, ])
+  }
+
+  openxlsx2::wb_save(wb, file = file, overwrite = TRUE)
 }
 
 #' @rdname affirm_report
 #' @export
-affirm_report_raw_data <- function(variable_labels = FALSE) {
+affirm_report_raw_data <- function() {
   .check_affirm_initialized()
   df_report <-
     get(x = "df_affirmations", envir = env_affirm_logs) |>
@@ -104,14 +105,7 @@ affirm_report_raw_data <- function(variable_labels = FALSE) {
     ) |>
     dplyr::arrange(.data$error_n == 0L, .data$priority, dplyr::desc(.data$error_rate))
 
-  # add the variable labels to the top row if requested
-  if (isTRUE(variable_labels)) {
-    df_report$data <- lapply(df_report$data, .add_row_of_column_labels)
-  }
-
   df_report
 }
-
-