+++ tests/testthat/test-affirm_report.R
@@ -1,3 +1,12 @@
+mtcars_modified <- mtcars |>
+  tibble::rownames_to_column(var = "car")
+
+attr(mtcars_modified$car, 'label') <- "Car model"
+attr(mtcars_modified$mpg, 'label') <- "Miles/(US) gallon"
+attr(mtcars_modified$cyl, 'label') <- "Number of cylinders"
+attr(mtcars_modified$disp, 'label') <- "Displacement (cu.in.)"
+
+
 test_that("affirm_report() works", {
 
   expect_error({
@@ -27,17 +36,24 @@ test_that("affirm_report() works", {
     NA
   )
 
-  # labels are added to report when requested
+  # names are added according to glue syntax
   expect_snapshot({
-    mtcars2 <- dplyr::as_tibble(mtcars)
-    attr(mtcars2$mpg, 'label') <- "MPG LABEL"
     affirm_init(replace = TRUE)
     affirm_true(
-      mtcars2,
+      mtcars,
       label = "leave it all, no actions",
-      condition = mpg > 33
+      condition = mpg > 33,
+      data_frames = "mtcars",
+      id = 1
     )
-    affirm_report_raw_data(variable_labels = TRUE)$data
+    affirm_true(
+      mtcars,
+      label = "No. cylinders must be 4 or 6",
+      condition = cyl %in% c(4, 6),
+      data_frames = "mtcars",
+      id = 2
+    )
+    affirm_report_raw_data()$data
   })
 })
 
@@ -65,15 +81,8 @@ test_that("affirm_report() works, but skip in CI", {
 })
 
 
-test_that("affirm_report_excel() details", {
+test_that("affirmation name details", {
 
-  mtcars_modified <- mtcars |>
-    tibble::rownames_to_column(var = "car")
-
-  attr(mtcars_modified$car, 'label') <- "Car model"
-  attr(mtcars_modified$mpg, 'label') <- "Miles/(US) gallon"
-  attr(mtcars_modified$cyl, 'label') <- "Number of cylinders"
-  attr(mtcars_modified$disp, 'label') <- "Displacement (cu.in.)"
 
   expect_error({
     affirm_init(replace = TRUE)
@@ -114,8 +123,8 @@ test_that("affirm_report_excel() details", {
       data_frames = "mtcars"
     );
     tmp_xlsx <- tempfile(fileext = ".xlsx")
-    affirm_report_excel(file = tmp_xlsx, sheet_name = "{data_frames} {id} {total_n}")
-    openxlsx::read.xlsx(tmp_xlsx, sheet = "mtcars 1 32")},
+    affirm_report_excel(file = tmp_xlsx, affirmation_name = "{data_frames} {id} {total_n}")
+    openxlsx2::read_xlsx(tmp_xlsx, sheet = "mtcars 1 32")},
     NA
   )
 
@@ -128,9 +137,9 @@ test_that("affirm_report_excel() details", {
       id = 1,
       data_frames = "mtcars"
     )
-    affirm_report_excel(file = tempfile(fileext = ".xlsx"), sheet_name = "{data_frames}{id}{label}moooooooorreeeeecharacters")
-    },
-    "At least one sheet name exceeds the allowed 31 characters."
+    affirm_report_excel(file = tempfile(fileext = ".xlsx"), affirmation_name = "{data_frames}{id}{label}moooooooorreeeeecharacters")
+  },
+  "At least one sheet name exceeds the allowed 31 characters"
   )
 
   expect_error({
@@ -142,10 +151,10 @@ test_that("affirm_report_excel() details", {
       id = 1,
       data_frames = "mtcars"
     )
-    affirm_report_excel(file = tempfile(fileext = ".xlsx"), sheet_name = "{data.frames}{id}")
+    affirm_report_excel(file = tempfile(fileext = ".xlsx"), affirmation_name = "{data.frames}{id}")
   },
-  "`sheet_name` glue syntax expects one of"
   )
+
 })
 
 test_that("affirmations with zero errors carried forward", {
@@ -173,9 +182,29 @@ test_that("affirmations with zero errors carried forward", {
       data_frames = "mtcars"
     )
     affirm_report_raw_data()[["data"]][[1]]
-    },
-    "data.frame"
+  },
+  "data.frame"
   )
 
 })
 
+
+test_that("excel report helpers", {
+
+  expect_equal(
+    .compute_col_width(mtcars_modified[, 1:4]),
+    c("car" = 22, "mpg" = 8, "cyl" = 8, "disp" = 8, "Notes" = 30)
+  )
+
+  expect_equal(
+    .retrieve_labels(mtcars_modified[, 1:5]),
+    data.frame(
+      "car" = "Car model",
+      "mpg" = "Miles/(US) gallon",
+      "cyl" = "Number of cylinders",
+      "disp" = "Displacement (cu.in.)",
+      "hp" = NA_character_
+      )
+  )
+
+})